增加获取指定作业测试样例的接口 测试样例的weight字段类型改为Integer AssignmentResponseDTO的id类型改为Long 获取附件统一返回AttachmentResponseDTO

+++ src/main/java/org/sustech/orion/service/impl/CodeServiceImpl.java
@@ -153,7 +153,7 @@ public class CodeServiceImpl implements CodeService {
 
                 boolean passed = result.isExecutionSuccess() &&
                         result.getOutput().trim().equals(testCase.getExpectedOutput().trim());
-                double weight = testCase.getWeight();
+                int weight = testCase.getWeight();
                 totalWeight += weight;
                 if (passed) {
                     totalScore += weight;
